+++ NEWS.md
@@ -1,5 +1,26 @@
 # scales (development version)
 
+* `manual_pal()` now always returns an unnamed colour vector, which is easy to
+  use with `ggplot2::discrete_scale()` (@yutannihilation, #284).
+
+# scales 1.1.1
+
+* `breaks_width()` now handles `difftime`/`hms` objects (@bhogan-mitre, #244).
+
+* `hue_pal()` now correctly inverts color palettes when `direction = -1` 
+  (@dpseidel, #252). 
+
+* Internal `precision()`, used when `accuracy = NULL`, now does a better
+  job when duplicate values are present (@teunbrand, #251). It also does a 
+  better job when there's a mix of finite and non-finite values (#257).
+
+* New `oob_keep()` to keep data outside range, allowing for zoom-limits when 
+  `oob_keep` is used as `oob` argument in scales. Existing out of bounds 
+  functions have been renamed with the `oob_`-prefix to indicate their role 
+  (@teunbrand, #255).
+
+* `ordinal_french()` gains `plural` and `gender` arguments (@stephLH, #256).
+
 # scales 1.1.0
 
 * Axis breaks and labels have a new naming scheme: functions that generate